Graph rehaul + many improvements
Completely re-hauled graphs. Code now takes into account time that each image was taken and includes it on graph. Normalization based off last pre-image (added a bit ago in error-handling branch but reiterating it here). graph X-axis is now  based on time instead of frames. Seaborn used to improve how the graph looks. Included piece of information to the graph that shows the difference between pre and post in terms of intensity. Also added a legend so viewer understands what each part of the graph means. improved documentation and fixes it capitalization to make it more professional.  Also closes plotting and pool functions even if an error has occurred to prevent system slowdowns due to memory usage. Fixed scales of normalized graphs so the Y-axis is constant no matter what to make it more understandable.